Fix some warnings under Mac OS X.

+++ source/UtilityFunctions.h
@@ -15,28 +15,28 @@ The following two functions are used to convert color
 inline void
 glFogf(GLenum pname, const QColor& c)
 {
-    const GLfloat color[] = { c.redF(), c.greenF(), c.blueF() };
+    const GLfloat color[] = { (GLfloat)c.redF(), (GLfloat)c.greenF(), (GLfloat)c.blueF() };
     glFogfv(pname, color);
 }
 
 inline void
 glLightf(GLenum light, GLenum pname, const QColor& c)
 {
-    const GLfloat color[] = { c.redF(), c.greenF(), c.blueF() };
+    const GLfloat color[] = { (GLfloat)c.redF(), (GLfloat)c.greenF(), (GLfloat)c.blueF() };
     glLightfv(light, pname, color);
 }
 
 inline void
 glMaterialf(GLenum face, GLenum pname, const QColor& c)
 {
-    const GLfloat color[] = { c.redF(), c.greenF(), c.blueF() };
+    const GLfloat color[] = { (GLfloat)c.redF(), (GLfloat)c.greenF(), (GLfloat)c.blueF() };
     glMaterialfv(face, pname, color);
 }
 
 inline void
 glTexEnvf(GLenum target, GLenum pname, const QColor& c)
 {
-    const GLfloat color[] = { c.redF(), c.greenF(), c.blueF() };
+    const GLfloat color[] = { (GLfloat)c.redF(), (GLfloat)c.greenF(), (GLfloat)c.blueF() };
     glTexEnvfv(target, pname, color);
 }
